moved "application" registry types inside "application" directory
