+++ fileformats/application/serialization.py
@@ -1,6 +1,5 @@
-from pathlib import Path
 import json
-from ..core import __version__, mark, DataType
+from ..core import mark, DataType
 from ..core.mixin import WithClassifiers
 from ..generic import File
 from ..core.exceptions import FormatMismatchError
